- Tweaks to allow the graph to match the visuals from MOS
