+++ PredixMobileUIKit/UIColor+Predix.swift
@@ -826,3 +826,10 @@ extension UIColor {
     }
 }
 
+extension UIColor {
+    var redValue: CGFloat{ return CIColor(color: self).red }
+    var greenValue: CGFloat{ return CIColor(color: self).green }
+    var blueValue: CGFloat{ return CIColor(color: self).blue }
+    var alphaValue: CGFloat{ return CIColor(color: self).alpha }
+}
+
